Add damage taken component and DamageApplySystem that takes DamageInstances and adds them to an entity's DamageTaken.

+++ space-dots/Assets/Scripts/Damage/DamageInstance.cs
@@ -3,7 +3,7 @@ using Unity.Mathematics;
 
 public struct DamageInstance : IComponentData
 {
-    public float DamageAmount;
+    public int DamageAmount;
     public Entity DamagedEntity;
 }
